x/servicer/client/cli/query_session_storage_for_spec.go
package cli

import (
	"context"

	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
	"github.com/lavanet/lava/x/servicer/types"
	"github.com/spf13/cobra"
)

func CmdListSessionStorageForSpec() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "list-session-storage-for-spec",
		Short: "list all sessionStorageForSpec",
		RunE: func(cmd *cobra.Command, args []string) error {
			clientCtx := client.GetClientContextFromCmd(cmd)

			pageReq, err := client.ReadPageRequest(cmd.Flags())
			if err != nil {
				return err
			}

			queryClient := types.NewQueryClient(clientCtx)

			params := &types.QueryAllSessionStorageForSpecRequest{
				Pagination: pageReq,
			}

			res, err := queryClient.SessionStorageForSpecAll(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}

func CmdShowSessionStorageForSpec() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "show-session-storage-for-spec [index]",
		Short: "shows a sessionStorageForSpec",
		Args:  cobra.ExactArgs(1),
		RunE: func(cmd *cobra.Command, args []string) (err error) {
			clientCtx := client.GetClientContextFromCmd(cmd)

			queryClient := types.NewQueryClient(clientCtx)

			argIndex := args[0]

			params := &types.QueryGetSessionStorageForSpecRequest{
				Index: argIndex,
			}

			res, err := queryClient.SessionStorageForSpec(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}
